feat(replay): Share performance instrumentation with tracing (#9296)
This streamlines web-vital & performance observer handling, by exposing
a new `addPerformanceInstrumentationHandler` method from
`@sentry-internal/tracing`.

This works similar to the instrumentation in utils, where the first time
you add instrumentation for a given type, it will add a performance
observer. And any further calls will just add more callbacks. This way,
we avoid having multiple of the same performance observers.

Furthermore, this also aligns the handling of LCP capturing for replay.
We used to do this separately, now we use the same data as for
performance.

Finally, while doing this I noticed that a whole bunch of performance
observer stuff we used to capture in Replay, was actually discarded 😬 so
no need to capture these anymore at all. (We can always add it back
later, if needed)

Some integration tests needed slight adjustments for this, probably due
to minor timing semantics. But I think all the changes are
good/"correct".

I _also_ got rid of the event deduplication in replay.

Closes https://github.com/getsentry/sentry-javascript/issues/9246