feat(browser): Allow collecting of pageload profiles (#9317)
Profiling is currently unable to profile page loads which creates a gap
in the product as performance instrumentation seemingly can (via
synthetic transactions)

In order to bridge this gap, the plan is to provide a small JS snippet
that users can insert into their document which will initialize and
store a reference to the pageload profile on the sentry carrier. By
doing so we can hook back into our finishTransaction codepath and send a
profile associated to a transaction.

Would love to hear early thoughts on this approach from SDK maintainers
before I start adding test coverage.