fix(replay): ensure `replay_id` is not added to DSC if session expired (#9359)
We noticed we still sometimes see "too long sessions", and it seems this
is because of errors being tagged with a replay long after the session
expired. After some digging, this may be because we do not check session
expiration when adding the `replay_id` to the DSC. This PR fixes this.