fix(angular-ivy): Adjust package entry points to support Angular 17 with SSR config (#9412)
Adjust the entry points of `@sentry/angular-ivy`'s
`package.json` to point directly to FESM2015 bundles (= bundled ESM2015
code) instead of the UMD bundles. This fixes an error when the old (no
longer supported) UMD bundles were picked up by Vite in Angular apps
with SSR config (#9376).

A proper long term fix to this is to bump to Angular 15 in this package
which we can only do in a new major.