feat(replay): Allow to treeshake & configure compression worker URL (#9409)
This PR does two things:

1. Allow to configure a `workerUrl` in replay config, which is expected
to be an URL of a self-hosted worker script.
a. Added an example worker script, which is a built version of the
pako-based compression worker
a. Users can basically host this file themselves and point to it in
`workerUrl`, as long as it is on the same origin as the website itself.
  a. We can eventually document this in docs
1. Allows to configure `__SENTRY_EXCLUDE_REPLAY_WORKER__` in your build
to strip the default included web worker. You can configure this if
you're disabling compression anyhow, or if you want to configure a
custom web worker as in the above step.

Fixes https://github.com/getsentry/sentry-javascript/issues/6739, and
allows to reduce bundle size further. Once merged/released we can also
add this to the bundler plugins `bundleSizeOptimizations` options.

Note that we _do not recommend_ to disable the web worker completely. We
only recommend to tree shake the worker code if you provide a custom
worker URL - else, replay payloads will not be compressed, resulting in
much larger payloads sent over the network, which is bad for your
applications performance.

Also note that when providing a custom worker, it is your own
responsibility to keep it up to date - we try to keep the worker
interface stable, and the worker is generally not updated often, but you
should still check regularly when updating the SDK if the example worker
has changed.

---------

Co-authored-by: Billy Vong <billyvg@users.noreply.github.com>