ref(replay): Use rrweb for slow click detection (#9408)
This refactors the slow/rage click detection to use rrweb events instead
of adding our own event listeners.

* We look for incremental snapshot events, and update the
`_lastMutation` and `_lastScroll` properties based on them.
* If a full snapshot happens, we interpret it as a mutation. this is not
necessarily correct, but we don't know any different. also this
shouldn't matter much because a full snapshot only happens in error
mode, in which case things that happened before the full checkout should
be discarded anyhow, so 🤷

Integration tests for this are still passing, so it seems to me to be a
good change. Hopefully this:

1. Decreases perf overhead by only using a single instrumentation
instead of multiple overlapping ones
2. Increases consistency by ensuring that whatever we see changing in
the replay will also be considered the same way for slow/rage click
detection.