fix(replay): Add `BODY_PARSE_ERROR` warning & time out fetch response load (#9622)
This is a tricky one 😬 Basically, it is possible that fetch returns a
readable stream that is ongoing. So if we do `await response.text()`
this will be pending forever, if a stream is ongoing.

I haven't found a way to really check that is the case and avoid parsing
this at all 🤔 So the best I could come up with was to instead add a
timeout of 500ms when we stop waiting for the fetch body.

This should at least unblock waiting on this, but it will still mean
that the response continues to be parsed client-side - I don't think
there is a way to abort this 🤔

Additionally, this also refactors this a bit so we add a new
`BODY_PARSE_ERROR` meta warning if the parsing of the body fails, for
whatever reason. we may also use this in the replay UI cc @ryan953
somehow?

"fixes" https://github.com/getsentry/sentry-javascript/issues/9616