feat(core): Allow to pass `mechanism` as event hint (#9590)
This allows to pass `mechanism` as an event hint. Also, we now allow to
pass `hint` as an alternative to `CaptureContext` to `captureException`
as second argument.

We have quite a lot of code where we fork a scope and add an event
processor, only to add a mechanism to the event.
Since this is a quite common pattern, I figured it makes more sense to
allow to pass a mechanism as an EventHint.

In addition, I noticed that while `hub.captureException()` takes an
event hint as second argument, in the core exported `captureException()`
we take a `CaptureContext` as second argument instead (for legacy
reasons).

In order to be able to pass a mechanism there as well, I updated the
method signature to allow _either_ a CaptureContext _or_ an EventHint. I
wrote some tests covering this to make sure that works - it's a bit
tricky since both can be POJOs, but no fields overlap so we are able to
parse this together.