feat(utils): Refactor `addInstrumentationHandler` to dedicated methods (#9542)
This is mostly an internal utility, but changes a bit. The previous
implementation for adding instrumentation had a few issues:

* It was impossible to tree shake. So even if some instrumentation is
never used, the whole code for it is always included. Splitting this up
opens up future improvements there as well.
* It was not type safe, and actually there _were_ some subtle issues in
the code because of this - things could fail in unexpected ways, as we
have been doing a lot of type casting from/to any etc.


This PR splits up `addInstrumentationHandler` into e.g.
`addFetchInstrumentationHandler` etc. methods, which improves the actual
type safety a lot, because currently we've been duplicating the types
everywhere, sometimes in slightly differing ways, leading to potential
issues. **We had a bunch of issues with xhr & fetch data in Replay
recently, so I figured it would finally be a good time to clean this up
so we can ensure that the data we get is consistent & reliable.**
Hopefully this change will prevent issues like we have/had in replay
from popping up in the future.

Some additional notes:
* For the dom instrumentation, I picked the name
`addClickKeypressInstrumentationHandler()`. I'm open for other naming
ideas, but I found `addDomInstrumentation` a bit too broad, as this
could mean a lot of things - and we are strictly only instrumenting
clicks and keypresses here. Another name (if we want to be a bit
broader) could be e.g. `addDomEventInstrumentation` or something like
this?
* I updated the type for the XHR data handler to make the method & url
required. This was previously optional (but not in some of the code,
where we expected this to always be), plus we also did not handle the
case that `url` can be a `URL` instead of a string, which we now
convert.
* On the XHR data, we have one field `args` that 1. is not really used
and 2. was actually "buggy", as we always returned the args, but for
`open` and `send` the args are very different things (but we pretended
it's always the "open" args). Now, this is actually always the method &
url, and I also deprecated this for v8.
* On the XHR data, we also had an unused/unset `data` field, which I
removed (this was not set anywhere)
* All old code should still work normally, but I deprecated the generic
`addInstrumentationHandler()` method, for removal in v8.
* I also took the opportunity and split the instrumentations into
dedicated files, previously this was one mega-file which made it harder
to reason about things.