fix(tracing): Filter out invalid resource sizes (#9641)
There's a bug in some browsers that attaches huge resource sizes that
are completely unrealistic. Here's an example in chromium:
https://bugs.chromium.org/p/chromium/issues/detail?id=1324812#c25

To get around this, we add a filter to enforce that resource sizes
should only be attached if the size is < 2147483647 bytes (size of
maximum value of integer in c/c++).