feat(astro): Automatically add Sentry middleware in Astro integration (#9532)
This change adds automatic registration of our Astro middleware. This is
possible since Astro 3.5.2 by [adding
middleware](https://docs.astro.build/en/reference/integrations-reference/#addmiddleware-option)
entry points in the astro integration's setup hook.

This is backwards compatible with previous Astro versions because we can
simply check if the `addMiddleware` function exists and only make use of
it if it does.