fix(astro): Configure sourcemap assets directory for Vercel adapter (#9665)
This PR adds support for uploading sourcemaps when using the
`@astrojs/vercel/*` adapters.

It seems that the adapter updates the config.outdir value too late for
our previous detection logic to
correctly detect the .vercel output directory. Furthermore, the adapter
for some reason first saves server
files to <root>/dist only to copy it later to <root>/.vercel/functions.
This copy command seems to happen too
late for when our sourcemaps plugin runs.

For these two reasons, we try to detect the used adapter (love that this
is much simpler in Astro than
Sveltekit) and adjust the assets glob to search for files in both,
`dist` and `.vercel`.