feat(astro): Add `assets` option to source maps upload options (#9668)
Adds the `assets` option to the Astro integration source maps
upload options.

It behaves just like the `assets` option of the Vite plugin, taking a
(array of) glob(s) and overriding the
default values. 
This came up in #9591 and I think it makes sense to give users with more
advanced/custom setups the ability to
override our defaults.