fix(astro): Avoid adding the Sentry Vite plugin in dev mode (#9688)
Add a guard to avoid adding the Sentry Vite plugin for source
maps upload in dev mode.

Reason: The vite plugin ships with a Sentry SDK. This SDK conflicts with
the Astro server-side SDK. For example for some reason, it skews up the
line numbers of stack frames. Not exactly sure why that's happening but
my best guess is that the plugin SDK somehow changes code (maybe we even
inject debugIds or release values or something like this) that mixes up
line numbers.

Anyway, I think it's generally more correct to not use the plugin in dev
mode.