fix(replay): Capture JSON XHR response bodies (#9623)
This is stupid, now that I figured it out. Basically, if you set
`xhr.responseType = 'json'`, it will force `xhr.response` to be a POJO -
which we can't parse right now.

We now handle this case specifically. 

This also adds a new `UNPARSEABLE_BODY_TYPE` meta warning if we are not
getting a body because it is not matching any of the known/parsed types.

Closes https://github.com/getsentry/sentry-javascript/issues/9339