ref: Use `getCurrentScope()`/`hub.getScope()` instead of `configureScope()` (#9846)
In preparation for
https://github.com/getsentry/sentry-javascript/issues/9841, as we want
to deprecate `configureScope`, refactor our own usage of this away.

I used a new transformer of sentry-migr8 for this, which did most of the
heavy lifting. I only needed to fix the block usage (which is more a
stylistic issue than an actual problem).

In some follow up, I'll also create a migr8 transform to refactor hub
usage away, to e.g. refactor `getCurrentHub().getScope()` to
`getCurrentScope()` etc. But for now this is OK I think.