feat(core): Update `withScope` to return callback return value (#9866)
To align this with OpenTelemetry and make some things possible that are
currently not easily doable without `pushScope` / `popScope`.

Noticed this because currently it's not easily possible to e.g. use
`withScope` in places like
[this](https://github.com/getsentry/sentry-javascript/pull/9862#discussion_r1427826506).



This should be backwards compatible because any code that previously
relied on this returning `void` should still work.