feat(core): Extract scope application to util (#9804)
This is a try to move the application of scope data to events out of the
scope class into a util function.

This changes the flow to be:

1. The scope has a `getScopeData()` method that returns data that should
be applied to events.
2. The `prepareEvent` method uses that to actually apply scope data to
the event etc.
3. This also makes it much easier to do the experimentation in
node-experimental
(https://github.com/getsentry/sentry-javascript/pull/9799) because we
only need to overwrite this, and can leave the rest of the event
processing the same.

I _think_ this makes sense but would appreciate some more eyes on this
as well. For me the separation makes also more sense, as e.g. the
application of event processors etc. should not really be tied to the
scope at all.

This is also is a first step to then allow us to add global scopes more
easily.

---------

Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>