build(size-limit): Consistently use webpack (#9910)
We've been using esbuild (default, from`@size-limit/preset-small-lib`)
for most jobs, but webpack for the one with custom config, which is a
bit inconsistent.

This now updates this to always use webpack, which should be more
consistent...

If we eventually get rid of the custom webpack config (e.g. when we
merge this or something like it:
https://github.com/getsentry/sentry-javascript/pull/9826), we can revert
this back to use esbuild everywhere.