feat(node): Rework ANR to use worker script via an integration (#9823)
This PR reworks Node ANR detection to use a worker thread. Workers are
usually started via a path to a source file but this can cause issues
when bundlers are used. Instead, the worker code is bundled and included
in the source as a base64 string which can be used to launch a worker
via a data URL. The base64 code comes in at around 45KB.

Positives 👍
- No extra processes
- Only 10-15MB memory overhead (vs at least 50MB for a child process)
- Uses inspector API so we can remove the websockets implementation
- Doesn't require special cases/setup for Electron main process
- No longer runs the app entry point again as the child process code
   - Closes some outstanding Electron ANR issues
- ANR becomes just an integration since we don't need to intercept app
execution in the child
   - Less confusing setup and less chance of running the app twice

Negatives 👎
- Minimum supported Node version for ANR detection becomes v16 because
Node 14 does not support data URLs for workers 😢

## Usage

```ts
import * as Sentry from '@sentry/node';

Sentry.init({
  dsn: 'https://public@dsn.ingest.sentry.io/1337',
  integrations: [new Sentry.Integrations.Anr({ captureStackTrace: true, anrThreshold: 200 })],
});
```