feat(core): Ensure `startSpan` & `startSpanManual` fork scope (#9955)
@lforst noticed that currently `startSpan()` and `startSpanManual` do
not fork the scope, which is also a slightly different behavior than it
is in OTEL.

This adjusts this to better align, so that we always fork a scope. We
also always leave the span on the forked scope, even after it was
finished.

In a follow up, we can thus also get rid of the `finish` callback arg
for `startSpanManual`, as users can/should simply call `span.end()`
themselves then.