feat: Add server runtime metrics aggregator (#9894)
## Implements

- [x] 10 Second Bucketing: SDKs are required to bucket into 10 second
intervals (rollup in seconds) which is the current lower bound of metric
accuracy.
- [x] Flush Shift: SDKs are required to shift the flush interval by
random() * rollup_in_seconds. That shift is determined once per startup
to create jittering.
- [ ] Force flush: an SDK is required to perform force flushing ahead of
scheduled time if the memory pressure is too high. There is no rule for
this other than that SDKs should be tracking abstract aggregation
complexity (eg: a counter only carries a single float, whereas a
distribution is a float per emission).

## Caveats 

- Force flush requires Node.js 14+ support (FinalizationRegistry). I
recommend leaving it after v8 release to make the implementation a lot
easier.

---------

Co-authored-by: Abhijeet Prasad <aprasad@sentry.io>