fix(core): Rethrow caught promise rejections in `startSpan`, `startSpanManual`, `trace` (#9958)
Our previous implementation of `startSpan` et.al. assumed
that because we were "cloning" the original promise, unhandled promise
rejections would still bubble up to the global `onunhandledrejection`
handler so that Sentry would catch the error.

However, we tried investigating this and found out that our "cloning"
mechanism didn't work correctly and because we already added a rejection
handler, the promise rejection would _not always_ make it to the global
handler.

After adding multiple integration tests, I further narrowed the buggy
behaviour down to a rather special case: The `startSpan` call is not `await`ed
which for some reason will lead to the SDK not catching the error.
Unless, we apply the fix in this PR.

This patch removes this cloning mechanism in favour of directly attaching a
rejection handler to the promise. In this handler, we rethrow the error
which should trigger the global handlers.