feat(node): Local variables via async inspector in node 19+ (#9962)
This PR creates a new `LocalVariables` integration that uses the async
inspector API.

Rather than create a huge messy integration that supports both the sync
(node 18) and async (node >= v19) APIs, I created a new integration and
wrapped both the sync and async integrations with user facing
integration that switches depending on node version.

The async API doesn't require the stacking of callbacks that risks stack
overflows and limits the number of frames we dare to evaluate. When we
tried wrapping the sync API with promises, memory was leaked at an
alarming rate!

The inspector APIs are not available on all builds of Node so we have to
lazily load it and catch any exceptions.

I've had to use `dynamicRequire` because webpack picks up `import()` and
reports missing dependency when bundling for older versions of node.