feat(node-experimental): Allow to pass base span options to trace methods (#10006)
This was brought up in Slack, currently you cannot pass other span
properties directly to otel `startSpan()`. This now adds `attributes`,
`kind` and `startTime` to better align with OTEL.

Note that I chose to omit `root` and `links` options, not sure if we
want to support those in none-OTEL environments, and this API should
(~~) be the same for node & non-node in v8.