feat(core): Deprecate session APIs on hub and add global replacements (#10054)
Deprecates and adds top-level replacements for

* `hub.startSession` -> `Sentry.startSession`
* `hub.endSession` -> `Sentry.endSession`
* `hub.captureException` -> `Sentry.captureSession`

These APIs were mostly used in browser and node `startSessionTracking`
functions which were called on `Sentry.init`. I updated these usages
with the global replacements and it seems integration tests are happy
with it.
For now, we not only put the session onto the isolation scope but also
onto the current scope to avoid changing event processing and stuff like
the ANR worker, as well as avoid "soft-breaking" users who use
`hub.(get|set)Session`.