ref(utils): Clean up timestamp calculation code (#10069)
This PR removes the usage of `dynamicRequire` from
`packages/utils/src/time.ts` and cleans up our timestamp code to be
simpler and reduce bundle size. Removing `dynamicRequire` means that we
no longer rely on `perf_hooks` for the `performance` API, but instead
try to grab it from `globalThis.performance`. `performance` was added to
the global in Node 16, which means we'll fallback to using `Date.now()`
in Node 8, 10, 12, 14 (and in v8 just Node 14). I think that is an
acceptable tradeoff, we just reduce accuracy for those versions.

This does not refactor `browserPerformanceTimeOrigin` code at the bottom
of the file, I want to come back and touch that in a follow up PR to
reduce the amount of changes made here.

I would appreciate reviews/opinions on this, I'm not 100% confident in
the changes.