feat(core): Allow to pass `scope` to `startSpan` APIs (#10076)
Also allow to pass this to `withScope()`, which actually aligns with
OTEL as well, where you can also do that. If you pass a scope there, it
will use this instead of forking a new one.

This should be the last thing needed to refactor some
`span.startChild()` occurrences - you can now store the scope you want
to fork off, and pass this into `startSpan` as needed.