feat(core): Deprecate span `name` and `description` (#10056)
Instead, users should use `spanToJSON(span).description`.

In reality, users should mostly not actually need this - most usages are
internal things, or tests, right now.

While at this, I also updated the `TraceContext` type to not have a
`description`. This started failing in a bunch of tests, because while
in theory the trace context has a `description` field, in praxis we
always used to set `description` to `undefined` for transactions,
leading to this never being set. The work in this PR unifies the
name/description handling between span and transaction, which lead to
`description` not being undefined for transactions anymore, leading to
this actually showing up in the trace context.