feat(core): Deprecate transaction metadata in favor of attributes (#10097)
This deprecates any usage of `metadata` on transactions.

The main usages we have are to set `sampleRate` and `source` in there.
These I replaced with semantic attributes. For backwards compatibility,
when creating the transaction event we still check the metadata there as
well.

Other usage of metadata (mostly around `request`) remains intact for
now, we need to replace this in v8 - e.g. put this on the isolation
scope, probably.

This is the first usage of [Semantic
Attributes](https://github.com/getsentry/rfcs/blob/main/text/0116-sentry-semantic-conventions.md)
in the SDK!

This replaces https://github.com/getsentry/sentry-javascript/pull/10041