feat(core): Deprecate `scope.getSpan()` & `scope.setSpan()`
These APIs are not compatible with OTEL tracing, and thus will not be public API anymore in v8.

In v8, we'll add new (?) methods to the scope that are "internal", as for non-node based SDKs we'll still need to keep the active span on the scope. But these should not be reflected publicly, and not in types' Scope.

note that we'll also need to make sure to use types' `Scope` for all callbacks etc., which we currently don't do.