feat(node): Make `getModuleFromFilename` compatible with ESM (#10061)
This removes usage of `require.main.filename` to bring better ESM
compatability.

An extra bonus in that now `createGetModuleFromFilename` is ESM
compatible, we can use it in the Anr worker when `appRootPath` is
supplied (ie. Electron only for now).

### ⚠️ Fingerprinting 
This may change `module` fingerprinting for ESM since previously when
`require.main.filename` was not available, `process.cwd()` was used.