feat(core): Streamline integration function results to be compatible (#10135)
For the v7 cycle, in order to make this easier to handle. in v8, we need
to remove the unused `setupOnce()` calls again.

It would be nicer if we could do away with this, but it is pretty hard
as the options for a client currently are types as `Integration[]`, and
we use these both for setting & getting, making it hard to change this
without breaking anything. Now, this is much easier (it basically just
works), with the small downsides of shipping a few more bytes - but we
can simply remove this in v8 then.