feat(core): Deprecate `Span.instrumenter` (#10139)
Deprecate the `instrumenter` field on the `Span` interface and
class. The field is not part of the Otel Span API and hence it can't
exist publicly on our spans either. We can replace field with an `instanceof` check
when setting `instrumenter` on the event in v8. This is breaking though,
so for now, we only deprecate it. Users shouldn't need to worry about
this field anyway.