feat(core): Deprecate `Span.transaction` in favor of `getRootSpan` (#10134)
Deprecate the `transaction` field on the `Span` interface and
class. Instead, we'll store the span hierarchy in a different format
external to the span class. In node-experimental, we use a WeakMap
referencing the parent span as a data structure which we might need to
generally do in v8. This, however, is a breaking change, so for now we
deprecate the field but use it in the replacement utility function for
v7.