fix(tracing-internal): Prefer `fetch` init headers over `fetch` input headers (#10176)
When a `fetch` call is made with a `Request` object as the
input/resource param (i.e. the first) and additionally provided headers
in the fetch init options (the second param), browsers ignore the headers 
from the request object but only add the headers from the init options 
when making the request.

Our `fetch` instrumentation had it the other way around, preferring
headers from the request object over the headers in the options, as
reported in #10172 . This patch fixes this behaviour by essentially turning
around the logic we previously had.