fix(core): Take user from current scope when starting a session (#10153)
When creating a session, we previously only read the user from the
isolation scope. This was problematic because the user could be set in
`initialScope` which will be applied to the _current_ scope on SDK init.
This would happen before the session is started when using auto session
tracking.

This PR now takes the user from the current scope if it's set and falls
back to the isolation scope should the user be set there. Added an
integration test to cover this regression.