feat(core): Add `client.getIntegrationByName()` (#10130)
And also deprecate both `getIntegration()` as well as
`getIntegrationById()` which is only on the baseclient, but not on the
client type, anyhow :grimace:.

Usage:

```ts
const replay = getClient().getIntegrationByName('Replay');
```

Or, if you want to have an easier time with types:

```ts
const replay = getClient().getIntegrationByName<Replay>('Replay');
```

## Why do we need this

In v8, integrations will not be classes anymore, so you cannot pass a
class definition anymore to `getIntegration()`. We also decided to
remove the static `id` field, so you cannot find an integration via that
way anymore.

`getIntegrationById()` was always just on the baseclient, so not
properly types anyhow. It is also an inconsistent/weird naming because
we will not have an `id` anymore for integrations at all.