ref(core): Add missing class field deprecations
These fields were alredy deprecated on the interface, this patch
only adds the deprecation annotation to the JSDoc of the class fields.
After testing with eslint, linters will correctly infer the deprecation
status. VSCode also shows the fields as deprecated in the JSDoc but
doesn't apply the strikethrough. So for extra clarity, let's also
deprecate the class fields here.