feat(core): Deprecate `hub.bindClient()` & `makeMain()` (#10188)
Instead, users should use `setCurrentClient()` & `client.init()`.

For now, we can't fully replace our own test usages of this everywhere,
as we need it to test a clean hub state. But for users, there should be
a reasonable path forward for 99% of cases. I added a docs page
outlining how the different scenarios will work in v8 (and already
should work now, too).