feat(core): Deprecate `Span.spanRecorder` (#10199)
Deprecate the `spanRecorder` field on the `Span` class. For now
there's no replacement yet. I'm not entirely sure how we're going to
keep track of spans within the SDK. #9972 shows an alternative for
`IdleSpan` which no longer seems to rely on the `spanRecorder`. Another
possibility might be to make the recorder private which might be enough,
too.