test(node): Test proxy server (#10156)
This PR adds an integration that tests the `proxy` transport option.

To do this I've added a `.withMockSentryServer()` feature to the test
runner.

When optionally enabled, a basic express server is started which listens
with an envelope endpoint. Test scenarios can then use the `SENTRY_DSN`
environment variable to set the correct DSN.

### Why a Sentry server when we have this new stdout transport for
testing?

We need to test the transport and assosiated things like the proxy
option which can't use stdout parsing.

### Why use the stdout transport at all if we can just use this server?

Resources when running tests in parallel? 

Because they use the same test runner, it's relatively simple to migrate
tests between the two, so once we have a full test suite we can test
this further.