fix(astro): Fix import path when using external init files with default path (#10214)
Fix a bug in the Astro SDK where specifying an
external `sentry.(client|server).config.js` in the default location file
would cause a build error.

Previously, we checked for a supported file extension and instead of
adding an import statement for the path to the file, we'd only add the
file extension. So
```js
import 'mjs';
// instead of 
import 'path/to/sentry.client.config.mjs'
```

Fix the filename lookup function and adds a test to cover
regressions.