feat(types): Add `SerializedEvent` interface (pre v8) (#10240)
Add a `SerializedEvent` interface which more accurately
represents what an event envelope item contains (i.e. the JSON structure
of a serialized `Event` object). Specifically, the `spans` field now
correctly only contains the attributes the POJO span will continue to
hold, i.e. without the deprecations and methods declared in the `Span`
interface.

In v8, we should change the `EventItem` type for the event envelope but
doing so now is a breaking change I believe (And arguably, not too
important to do now anyway).