feat: Refactor exposed `defaultIntegrations` to `getDefaultIntegrations()` (#10243)
The current implementation has two problems:

1. It is weird that you can accidentally mutate the default integrations
of another package
2. We sometimes have logic-based default integrations - e.g. adding an
integration only if tracing is enabled, or similar. This means that
either we have to add some logic in the _upstream_ SDK to ensure this is
still added even if downstream SDKs overwrite default integrations, or
we need to duplicate the logic in the _downstream_ SDKs. With this new
method, we can instead centralize this, and downstream SDKs simply need
to call upstream `getDefaultIntegrations(options)`.