fix(next): Fix custom integrations (#10220)
The usage of this was not really working well to begin with, and even
worse with the new functional integrations. Because if the user adds the
integration themselves (e.g. `integrations: [new RewriteFrames()]`), it
will not actually get the correct iteratee at all.

Overall it is much cleaner anyhow to just fork the integrations properly
and use them instead of the default one - then we can rely on the
standard behavior of merging integrations etc.

We need to do the same for basically all usages of
`addOrUpdateIntegration`, as that actually does not work at all anymore
with the functional integrations 😬 (and in many instances never really
worked properly if users passed in a custom integration themselves).

Co-authored-by: Abhijeet Prasad <aprasad@sentry.io>