fix(wasm): Add `@sentry/core` as a dependency (#10283)
By using functional integrations in #10230, we started importing from
`@sentry/core` in the WASM integration. However, we didn't register
`@sentry/core` as a dependency, making rollup bundle core into the
package output. This changed the `build/npm` directory structure, making
our entry points in `package.json` invalid.

This PR fixes things by simply registering core as a dependency of wasm.
If we move WASM to core (which I strongly think we should do), we'll be
able to get rid of this again.