feat(core): Expose `isInitialized()` to replace checking via `getClient`
Currently, you can use `Sentry.getClient() !== undefiend` to check if Sentry was initialized. In v8, we want to change this so that this _always_ returns a client (possibly a Noop client), so this check will not work anymore there. Instead, we can provide a new util that does this explicitly, where we can control what it checks under the hood.