feat(replay): Deprecate Replay, ReplayCanvas, Feedback classes (#10270)
Instead, users should use the new functional styles.

Note that we'll probably actually un-deprecate `Replay` in some form in
v8, as we'll be keeping the class around there for sure (as there is a
lot of logic in there...). But users should not use it, so deprecating
this now!

While at it, I also deprecated the old `InitSentryForEmber` method in
favor of `init()`.

It's slightly unfortunate that I missed this, but we probably shouldn't
have exposed `ReplayCanvas` as a class anymore at all 😬 maybe we wait
before we document this etc. until we merged the functional style. cc
@billyvg